+++ tests/PHPUnit/Integration/Plugin/Dimension/VisitDimensionTest.php
@@ -6,13 +6,13 @@
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
 
-namespace Piwik\Plugins\Test;
 // there is a test that requires the class to be defined in a plugin
+namespace Piwik\Plugins\Test;
 
 use Piwik\Plugin\Dimension\VisitDimension;
 use Piwik\Plugin\Segment;
 use Piwik\Plugin\Manager;
-use Piwik\Tests\Framework\TestCase\UnitTestCase;
+use Piwik\Tests\Framework\TestCase\IntegrationTestCase;
 use Piwik\Tracker\Request;
 use Piwik\Tracker\Visitor;
 
@@ -62,7 +62,7 @@ class FakeConversionVisitDimension extends FakeVisitDimension
 /**
  * @group Core
  */
-class Plugin_VisitDimensionTest extends UnitTestCase
+class VisitDimensionTest extends IntegrationTestCase
 {
     /**
      * @var FakeVisitDimension
